STYLE: startpoint 타입 float -> double로 변경

+++ backend/src/main/java/com/showing/backend/api/response/PerformanceDetailRes.java
@@ -26,7 +26,7 @@ public class PerformanceDetailRes {
     int performanceType;
 
     @ApiModelProperty(name = "공연 평균 별점", example = "1")
-    float starPointAverage;
+    double starPointAverage;
 
     @ApiModelProperty(name = "공연 최근 시즌 정보")
     SeasonRes seasonRes;
@@ -34,7 +34,7 @@ public class PerformanceDetailRes {
     @ApiModelProperty(name = "공연 리뷰 목록 미리보기")
     List<PreviewReviewByPerformanceRes> previewReviewList;
 
-    public static PerformanceDetailRes of(Performance performance, float starPointAvg, SeasonRes seasonRes, List<PreviewReviewByPerformanceRes> previewReviewList){
+    public static PerformanceDetailRes of(Performance performance, double starPointAvg, SeasonRes seasonRes, List<PreviewReviewByPerformanceRes> previewReviewList){
         PerformanceDetailRes res = new PerformanceDetailRes();
         res.setPerformanceId(performance.getId());
         res.setPerformanceName(performance.getPerformanceName());
